# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationPrefOut": {
        "properties": {
          "events": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Events",
            "type": "object"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "muted": {
            "title": "Muted",
            "type": "boolean"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "label",
          "muted",
          "events"
        ],
        "title": "NotificationPrefOut",
        "type": "object"
      },
      "NotificationPrefUpdate": {
        "properties": {
          "events": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events"
          },
          "muted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Muted"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "scope"
        ],
        "title": "NotificationPrefUpdate",
        "type": "object"
      },
      "NotificationPrefsOut": {
        "properties": {
          "event_types": {
            "items": {
              "type": "string"
            },
            "title": "Event Types",
            "type": "array"
          },
          "prefs": {
            "items": {
              "$ref": "#/components/schemas/NotificationPrefOut"
            },
            "title": "Prefs",
            "type": "array"
          }
        },
        "required": [
          "event_types",
          "prefs"
        ],
        "title": "NotificationPrefsOut",
        "type": "object"
      },
...
      "PlayerFollowIn": {
        "description": "Body of ``PUT /players/follows/{athlete_id}``.\n\nv1 constraint: ``team_id`` must be a followed team and the athlete\nmust be on its stored roster (the roster is the picker).",
        "properties": {
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "team_id": {
            "title": "Team Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "team_id",
          "league_id"
        ],
        "title": "PlayerFollowIn",
        "type": "object"
      },
      "PlayerFollowOut": {
        "description": "One followed individual player (on a followed team-sport team).",
        "properties": {
          "athlete_id": {
            "title": "Athlete Id",
            "type": "string"
          },
          "followed_at": {
            "format": "date-time",
            "title": "Followed At",
            "type": "string"
          },
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "team_id": {
            "title": "Team Id",
            "type": "string"
          }
        },
        "required": [
          "athlete_id",
          "name",
          "team_id",
          "league_id",
          "followed_at"
        ],
        "title": "PlayerFollowOut",
        "type": "object"
      },
...
      "StatLeaderOut": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "followed": {
            "default": false,
            "title": "Followed",
            "type": "boolean"
          },
          "highlighted": {
            "default": false,
            "title": "Highlighted",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "stat_label": {
            "title": "Stat Label",
            "type": "string"
          },
          "team_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Color"
          },
          "team_id": {
            "default": "",
            "title": "Team Id",
            "type": "string"
          },
          "team_logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Logo Url"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "rank",
          "player_id",
          "name",
          "team_name",
          "value",
          "stat_label",
          "detail"
        ],
        "title": "StatLeaderOut",
        "type": "object"
      },
      "StatLeadersOut": {
        "properties": {
          "league_id": {
            "title": "League Id",
            "type": "string"
          },
          "league_name": {
            "title": "League Name",
            "type": "string"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StatLeaderOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "sport": {
            "title": "Sport",
            "type": "string"
          },
          "stat_label": {
            "default": "",
            "title": "Stat Label",
            "type": "string"
          }
        },
        "required": [
          "league_id",
          "league_name",
          "sport"
        ],
        "title": "StatLeadersOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SportsDash",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/leaders/{league_id}": {
      "get": {
        "operationId": "leaders_api_leaders__league_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "league_id",
            "required": true,
            "schema": {
              "title": "League Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatLeadersOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Leaders"
      }
    },
...
    "/api/notifications/prefs": {
      "get": {
        "operationId": "get_notification_prefs_api_notifications_prefs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPrefsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Notification Prefs"
      },
      "put": {
        "description": "Merge one scope's preferences, commit, return the full refreshed set.",
        "operationId": "put_notification_pref_api_notifications_prefs_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationPrefUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPrefsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Notification Pref"
      }
    },
...
    "/api/players/follows/{athlete_id}": {
      "delete": {
        "operationId": "unfollow_player_api_players_follows__athlete_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "athlete_id",
            "required": true,
            "schema": {
              "title": "Athlete Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unfollow Player"
      },
      "put": {
        "description": "Follow (or re-follow) one player; validates the v1 roster constraint.\n\nESPN athlete ids are only unique per sport (the same bare id can name\nboth an NHL and an MLB athlete), and follows key on the bare id alone\n\u2014 so an id already followed under a DIFFERENT league is rejected with\n409 rather than silently overwriting that follow.  The roster row's\ncurrent status seeds the follow's alert baseline, so following an\nalready-injured player fires no alert.",
        "operationId": "follow_player_api_players_follows__athlete_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "athlete_id",
            "required": true,
            "schema": {
              "title": "Athlete Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerFollowIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerFollowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Follow Player"
      }
    },
...
  }
}